Added memory objects list

+++ VulkanSimplifiedStatic/Include/VSCommon/VSMemoryTypeData.cpp
@@ -5,7 +5,7 @@ namespace VulkanSimplified
 {
 	MemoryTypeData::MemoryTypeData()
 	{
-		memoryIndex = std::numeric_limits<size_t>::max();
+		memoryIndex = std::numeric_limits<uint32_t>::max();
 		properties = std::numeric_limits<MemoryTypeProperties>::max();
 	}
 }
